bot/requirements.txt
google-generativeai
gitpython
pytest
